# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AutoPeace API — AI-powered conflict forecasting and peace research
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/config:
    get:
      operationId: getAdminConfig
      tags: [admin]
      summary: Get admin configuration
      security:
        - adminKey: []
      responses:
        "200":
          description: Admin configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminConfigResponse"
        "401":
          description: Unauthorized
    post:
      operationId: updateAdminConfig
      tags: [admin]
      summary: Update admin configuration
      security:
        - adminKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminConfigUpdate"
      responses:
        "200":
          description: Updated configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminConfigResponse"
        "401":
          description: Unauthorized
  /admin/run:
    post:
      operationId: triggerRun
      tags: [admin]
      summary: Trigger an immediate autoresearch cycle
      security:
        - adminKey: []
      responses:
        "200":
          description: Run triggered
          content:
            application/json:
              schema:
                type: object
                properties:
                  cycleId:
                    type: string
                  message:
                    type: string
                required: [cycleId, message]
        "401":
          description: Unauthorized
        "409":
          description: Already running
  /admin/sources:
    get:
      operationId: listEvidenceSources
      tags: [admin]
      summary: List evidence sources
      security:
        - adminKey: []
      responses:
        "200":
          description: Evidence sources
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/EvidenceSource"
                required: [data]
        "401":
          description: Unauthorized
  /admin/sources/{id}:
    patch:
      operationId: updateEvidenceSource
      tags: [admin]
      summary: Update an evidence source
      security:
        - adminKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvidenceSourceUpdate"
      responses:
        "200":
          description: Updated source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceSource"
        "401":
          description: Unauthorized
        "404":
          description: Not found
  /admin/costs-summary:
    get:
      operationId: getAdminCostsSummary
      tags: [admin]
      summary: Get LLM provider cost summary
      security:
        - adminKey: []
      responses:
        "200":
          description: Cost summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CostSummary"
        "401":
          description: Unauthorized
  /admin/deal-run:
    post:
      operationId: triggerDealRun
      tags: [admin]
      summary: Trigger an immediate deal autoresearch cycle (Task B)
      security:
        - adminKey: []
      responses:
        "200":
          description: Deal cycle started
          content:
            application/json:
              schema:
                type: object
                properties:
                  cycleId:
                    type: string
                  message:
                    type: string
                required: [cycleId, message]
        "401":
          description: Unauthorized
        "409":
          description: Already running
  /admin/deal-cycles:
    get:
      operationId: listDealCycles
      tags: [admin]
      summary: List recent deal autoresearch cycles with status
      security:
        - adminKey: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of deal cycles
          content:
            application/json:
              schema:
                type: object
                properties:
                  cycles:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        status:
                          type: string
                          enum: [running, complete, error]
                        startedAt:
                          type: string
                          format: date-time
                        completedAt:
                          type: string
                          format: date-time
                          nullable: true
                        dealId:
                          type: string
                          nullable: true
                        error:
                          type: string
                          nullable: true
                      required: [id, status, startedAt]
                required: [cycles]
        "401":
          description: Unauthorized
...
  /admin/pipeline/config:
    get:
      operationId: getPipelineConfig
      tags: [admin]
      summary: Get the current pipeline stage-to-provider-model assignment (admin only)
      security:
        - adminKey: []
      responses:
        "200":
          description: Pipeline configuration
          content:
            application/json:
              schema:
                type: object
                properties:
                  stages:
                    type: array
                    items:
                      type: object
                      properties:
                        stage: { type: integer }
                        name: { type: string }
                        role: { type: string }
                        provider: { type: string, enum: [anthropic, openai, gemini] }
                        model: { type: string }
                      required: [stage, name, role, provider, model]
                  constraint: { type: string }
                required: [stages, constraint]
        "401":
          description: Unauthorized
...
components:
  securitySchemes:
    adminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
  schemas:
...
    AdminConfigResponse:
      type: object
      properties:
        cadence:
          type: string
          enum: [hourly, daily, weekly, manual]
        budgetCapUsd:
          type: number
        isPaused:
          type: boolean
        anthropicModel:
          type: string
        openaiModel:
          type: string
        geminiModel:
          type: string
        generationProvider:
          type: string
          enum: [anthropic, openai, gemini]
        generationModel:
          type: string
        evaluationProvider:
          type: string
          enum: [anthropic, openai, gemini]
        evaluationModel:
          type: string
        adversarialProvider:
          type: string
          enum: [anthropic, openai, gemini]
        adversarialModel:
          type: string
        forecastingProvider:
          type: string
          enum: [anthropic, openai, gemini]
        forecastingModel:
          type: string
        extractionProvider:
          type: string
          enum: [anthropic, openai, gemini]
        extractionModel:
          type: string
        judgePanelAnthropicModel:
          type: string
        judgePanelOpenaiModel:
          type: string
        judgePanelGeminiModel:
          type: string
        stage1Provider:
          type: string
          enum: [anthropic, openai, gemini]
        stage1Model:
          type: string
        stage2Provider:
          type: string
          enum: [anthropic, openai, gemini]
        stage2Model:
          type: string
        stage3Provider:
          type: string
          enum: [anthropic, openai, gemini]
        stage3Model:
          type: string
        stage4Provider:
          type: string
          enum: [anthropic, openai, gemini]
        stage4Model:
          type: string
        stage5Provider:
          type: string
          enum: [anthropic, openai, gemini]
        stage5Model:
          type: string
        stage6Provider:
          type: string
          enum: [anthropic, openai, gemini]
        stage6Model:
          type: string
        stage7Provider:
          type: string
          enum: [anthropic, openai, gemini]
        stage7Model:
          type: string
        stage8Provider:
          type: string
          enum: [anthropic, openai, gemini]
        stage8Model:
          type: string
      required:
        - cadence
        - budgetCapUsd
        - isPaused
        - anthropicModel
        - openaiModel
        - geminiModel
        - generationProvider
        - generationModel
        - evaluationProvider
        - evaluationModel
        - adversarialProvider
        - adversarialModel
        - forecastingProvider
        - forecastingModel
        - extractionProvider
        - extractionModel
    AdminConfigUpdate:
      type: object
      properties:
        cadence:
          type: string
          enum: [hourly, daily, weekly, manual]
        budgetCapUsd:
          type: number
        isPaused:
          type: boolean
        anthropicModel:
          type: string
        openaiModel:
          type: string
        geminiModel:
          type: string
        generationProvider:
          type: string
          enum: [anthropic, openai, gemini]
        generationModel:
          type: string
        evaluationProvider:
          type: string
          enum: [anthropic, openai, gemini]
        evaluationModel:
          type: string
        adversarialProvider:
          type: string
          enum: [anthropic, openai, gemini]
        adversarialModel:
          type: string
        forecastingProvider:
          type: string
          enum: [anthropic, openai, gemini]
        forecastingModel:
          type: string
        extractionProvider:
          type: string
          enum: [anthropic, openai, gemini]
        extractionModel:
          type: string
        judgePanelAnthropicModel:
          type: string
        judgePanelOpenaiModel:
          type: string
        judgePanelGeminiModel:
          type: string
        stage1Provider:
          type: string
          enum: [anthropic, openai, gemini]
        stage1Model:
          type: string
        stage2Provider:
          type: string
          enum: [anthropic, openai, gemini]
        stage2Model:
          type: string
        stage3Provider:
          type: string
          enum: [anthropic, openai, gemini]
        stage3Model:
          type: string
        stage4Provider:
          type: string
          enum: [anthropic, openai, gemini]
        stage4Model:
          type: string
        stage5Provider:
          type: string
          enum: [anthropic, openai, gemini]
        stage5Model:
          type: string
        stage6Provider:
          type: string
          enum: [anthropic, openai, gemini]
        stage6Model:
          type: string
        stage7Provider:
          type: string
          enum: [anthropic, openai, gemini]
        stage7Model:
          type: string
        stage8Provider:
          type: string
          enum: [anthropic, openai, gemini]
        stage8Model:
          type: string
    EvidenceSource:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        url:
          type: string
        isEnabled:
          type: boolean
        lastFetchedAt:
          type: string
          format: date-time
          nullable: true
        fetchFrequencyMinutes:
          type: integer
      required:
        - id
        - name
        - type
        - url
        - isEnabled
        - fetchFrequencyMinutes
    EvidenceSourceUpdate:
      type: object
      properties:
        isEnabled:
          type: boolean
        fetchFrequencyMinutes:
          type: integer
    CostSummary:
      type: object
      properties:
        totalCostUsd:
          type: number
        totalTokens:
          type: integer
        byProvider:
          type: object
          additionalProperties:
            type: object
            properties:
              costUsd:
                type: number
              tokens:
                type: integer
            required:
              - costUsd
              - tokens
        byCycle:
          type: array
          items:
            type: object
            properties:
              cycleId:
                type: string
              startedAt:
                type: string
                format: date-time
              costUsd:
                type: number
              tokens:
                type: integer
            required:
              - cycleId
              - startedAt
              - costUsd
              - tokens
      required:
        - totalCostUsd
        - totalTokens
        - byProvider
        - byCycle
...