# lib/api-zod/src/generated/types/adminConfigUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AutoPeace API — AI-powered conflict forecasting and peace research
 * OpenAPI spec version: 0.1.0
 */
import type { AdminConfigUpdateAdversarialProvider } from "./adminConfigUpdateAdversarialProvider";
import type { AdminConfigUpdateCadence } from "./adminConfigUpdateCadence";
import type { AdminConfigUpdateEvaluationProvider } from "./adminConfigUpdateEvaluationProvider";
import type { AdminConfigUpdateExtractionProvider } from "./adminConfigUpdateExtractionProvider";
import type { AdminConfigUpdateForecastingProvider } from "./adminConfigUpdateForecastingProvider";
import type { AdminConfigUpdateGenerationProvider } from "./adminConfigUpdateGenerationProvider";
import type { AdminConfigUpdateStage1Provider } from "./adminConfigUpdateStage1Provider";
import type { AdminConfigUpdateStage2Provider } from "./adminConfigUpdateStage2Provider";
import type { AdminConfigUpdateStage3Provider } from "./adminConfigUpdateStage3Provider";
import type { AdminConfigUpdateStage4Provider } from "./adminConfigUpdateStage4Provider";
import type { AdminConfigUpdateStage5Provider } from "./adminConfigUpdateStage5Provider";
import type { AdminConfigUpdateStage6Provider } from "./adminConfigUpdateStage6Provider";
import type { AdminConfigUpdateStage7Provider } from "./adminConfigUpdateStage7Provider";
import type { AdminConfigUpdateStage8Provider } from "./adminConfigUpdateStage8Provider";

export interface AdminConfigUpdate {
  cadence?: AdminConfigUpdateCadence;
  budgetCapUsd?: number;
  isPaused?: boolean;
  anthropicModel?: string;
  openaiModel?: string;
  geminiModel?: string;
  generationProvider?: AdminConfigUpdateGenerationProvider;
  generationModel?: string;
  evaluationProvider?: AdminConfigUpdateEvaluationProvider;
  evaluationModel?: string;
  adversarialProvider?: AdminConfigUpdateAdversarialProvider;
  adversarialModel?: string;
  forecastingProvider?: AdminConfigUpdateForecastingProvider;
  forecastingModel?: string;
  extractionProvider?: AdminConfigUpdateExtractionProvider;
  extractionModel?: string;
  judgePanelAnthropicModel?: string;
  judgePanelOpenaiModel?: string;
  judgePanelGeminiModel?: string;
  stage1Provider?: AdminConfigUpdateStage1Provider;
  stage1Model?: string;
  stage2Provider?: AdminConfigUpdateStage2Provider;
  stage2Model?: string;
  stage3Provider?: AdminConfigUpdateStage3Provider;
  stage3Model?: string;
  stage4Provider?: AdminConfigUpdateStage4Provider;
  stage4Model?: string;
  stage5Provider?: AdminConfigUpdateStage5Provider;
  stage5Model?: string;
  stage6Provider?: AdminConfigUpdateStage6Provider;
  stage6Model?: string;
  stage7Provider?: AdminConfigUpdateStage7Provider;
  stage7Model?: string;
  stage8Provider?: AdminConfigUpdateStage8Provider;
  stage8Model?: string;
}